update chat dan return

+++ app/Http/Controllers/Admin/TransaksiController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Controllers\RajaOngkirController;
 use App\Mail\Pesanan;
+use App\Models\Tracking;
 use App\Models\TrackingDetails;
 use App\Models\Transaksi;
 use Illuminate\Http\Request;
@@ -42,11 +43,13 @@ class TransaksiController extends Controller
             ->where('kodeTrx', $kodeTrx)
             ->where('status_bayar', 'berhasil')
             ->first();
-        foreach ($trx->trackings->trackings_details as $trackLama) {
-            $trackLama->delete();
+        if ($trx->trackings) {
+            foreach ($trx->trackings->trackings_details as $trackLama) {
+                $trackLama->delete();
+            }
         }
         $rajaongkir = new RajaOngkirController();
-        $respon = $rajaongkir->cekResi($request->resi, $request->layanan, $trx->trackings->last_phone);
+        $respon = $rajaongkir->cekResi($request->resi, $request->layanan, $trx->trackings->last_phone ?? substr($trx->transaksi_details->no_penerima, -5));
         // dd($respon);
         if (!$respon['data']) {
             $trx->trackings->update([
@@ -57,11 +60,21 @@ class TransaksiController extends Controller
             Alert::error('Error: ' . $respon['meta']['message']);
             return back()->withInput();
         }
-        $trx->trackings->update([
-            'resi' => $request->resi,
-            'ekspedisi' => $request->layanan,
-            'status' => 'dalam pengiriman',
-        ]);
+        if (!$trx->trackings) {
+            Tracking::create([
+                'transaksi_id' => $trx->id,
+                'resi' => $request->resi,
+                'ekspedisi' => $request->layanan,
+                'status' => 'dalam pengiriman',
+                'last_phone' => substr($trx->transaksi_details->no_penerima, -5),
+            ]);
+        } else {
+            $trx->trackings->update([
+                'resi' => $request->resi,
+                'ekspedisi' => $request->layanan,
+                'status' => 'dalam pengiriman',
+            ]);
+        }
         if ($respon['data']['delivered']) {
             if (isset($respon['data']['delivery_status'])) {
                 if (str_contains($respon['data']['delivery_status']['pod_date'], ' ')) {
@@ -78,9 +91,6 @@ class TransaksiController extends Controller
                 'received_at' => $datetime,
             ]);
         }
-        // $datas = file_get_contents(public_path('assets/contoh.json'));
-        // $datas = json_decode($datas, true);
-        // foreach ($datas['data']['manifest'] as $data) {
         foreach ($respon['data']['manifest'] as $data) {
             if (isset($data['manifest_date'])) {
                 if (str_contains($data['manifest_date'], ' ')) {
